requirements.txt
requests>=2.28.2
pandas>=1.5.3
python-dotenv>=1.0.0
gspread>=5.7.0
oauth2client>=4.1.3
gspread-dataframe>=3.3.0
pytz>=2023.3